requirements.txt
asgiref 
async
Babel 
certifi 
click 
colorama 
cycler 
defusedxml 
distlib 
filelock 
Flask 
Flask
fonttools 
greenlet 
itsdangerous 
Jinja2 
joblib 
kiwisolver 
lz4 
mariadb 
MarkupSafe 
matplotlib 
mock 
mysqlclient 
numpy 
olefile 
packaging 
pandas 
pathtools 
Pillow 
platformdirs 
pyparsing 
pytz 
PyYAML 
scipy 
seaborn 
six 
SQLAlchemy 
threadpoolctl 
tornado 
typing_extensions 
virtualenv 
watchdog 
Werkzeug 
scikit-learn
flask_sqlalchemy
python-dotenv
flask-serialize
flask-jwt-extended
flask-cors
openpyxl